+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## **1.2.0** (in progress)
 * Update eslint to v6.2.1 (UIPFIMP-8)
+* Fix wording in relink modal (UIPFIMP-10)
 
 ## [1.1.2](https://github.com/folio-org/ui-plugin-find-import-profile/tree/v1.1.2) (2020-01-03)